mainFile/requirements.txt
aiogram
python-dotenv
sqlalchemy
asyncpg
openai
requests
